refactor string utils for better consistency
