+++ std/assembly/util/string.ts
@@ -139,3 +139,73 @@ export function parse<T>(str: string, radix: i32 = 0): T {
   // @ts-ignore: type
   return sign * num;
 }
+
+// FIXME: naive implementation
+export function parseFloat(str: String): f64 {
+  var len: i32 = str.length;
+  if (!len) return NaN;
+
+  var ptr = changetype<usize>(str);
+  var code = <i32>load<u16>(ptr);
+
+  // determine sign
+  var sign: f64;
+  // trim white spaces
+  while (isWhiteSpaceOrLineTerminator(code)) {
+    code = <i32>load<u16>(ptr += 2);
+    --len;
+  }
+  if (code == CharCode.MINUS) {
+    if (!--len) return NaN;
+    code = <i32>load<u16>(ptr += 2);
+    sign = -1;
+  } else if (code == CharCode.PLUS) {
+    if (!--len) return NaN;
+    code = <i32>load<u16>(ptr += 2);
+    sign = 1;
+  } else {
+    sign = 1;
+  }
+
+  // calculate value
+  var num: f64 = 0;
+  while (len--) {
+    code = <i32>load<u16>(ptr);
+    if (code == CharCode.DOT) {
+      ptr += 2;
+      let fac: f64 = 0.1; // precision :(
+      while (len--) {
+        code = <i32>load<u16>(ptr);
+        if (code == CharCode.E || code == CharCode.e) {
+          assert(false); // TODO
+        }
+        code -= CharCode._0;
+        if (<u32>code > 9) break;
+        num += <f64>code * fac;
+        fac *= 0.1;
+        ptr += 2;
+      }
+      break;
+    }
+    code -= CharCode._0;
+    if (<u32>code >= 10) break;
+    num = (num * 10) + code;
+    ptr += 2;
+  }
+  return sign * num;
+}
+
+export function parseInt(str: String, radix: i32 = 0): f64 {
+  // @ts-ignore: string <-> String
+  return parse<f64>(str, radix);
+}
+
+export function parseI32(str: String, radix: i32 = 0): i32 {
+  // @ts-ignore: string <-> String
+  return parse<i32>(str, radix);
+}
+
+export function parseI64(str: String, radix: i32 = 0): i64 {
+  // @ts-ignore: string <-> String
+  return parse<i64>(str, radix);
+}
