+++ tests/compiler/number.untouched.wat
@@ -3376,7 +3376,7 @@
   if
    i32.const 0
    i32.const 1736
-   i32.const 196
+   i32.const 198
    i32.const 4
    call $~lib/builtins/abort
    unreachable
