+++ tests/compiler/std/array.optimized.wat
@@ -7243,7 +7243,7 @@
   if
    i32.const 0
    i32.const 4360
-   i32.const 196
+   i32.const 198
    i32.const 4
    call $~lib/builtins/abort
    unreachable
