+++ tests/compiler/std/array.untouched.wat
@@ -11392,7 +11392,7 @@
   if
    i32.const 0
    i32.const 4360
-   i32.const 40
+   i32.const 42
    i32.const 4
    call $~lib/builtins/abort
    unreachable
@@ -12165,7 +12165,7 @@
   if
    i32.const 0
    i32.const 4360
-   i32.const 196
+   i32.const 198
    i32.const 4
    call $~lib/builtins/abort
    unreachable
