+++ tests/compiler/std/string-utf8.untouched.wat
@@ -1951,7 +1951,7 @@
        if
         i32.const 0
         i32.const 128
-        i32.const 450
+        i32.const 452
         i32.const 8
         call $~lib/builtins/abort
         unreachable
@@ -2004,7 +2004,7 @@
         if
          i32.const 0
          i32.const 128
-         i32.const 454
+         i32.const 456
          i32.const 8
          call $~lib/builtins/abort
          unreachable
@@ -2099,7 +2099,7 @@
         if
          i32.const 0
          i32.const 128
-         i32.const 466
+         i32.const 468
          i32.const 8
          call $~lib/builtins/abort
          unreachable
@@ -2161,7 +2161,7 @@
   if
    i32.const 0
    i32.const 128
-   i32.const 475
+   i32.const 477
    i32.const 4
    call $~lib/builtins/abort
    unreachable
